pyproject.toml
[project]
name = "beyond-chatgpt"
version = "0.1.0"
description = "gpt-like-a-dev"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115.12",
    "jupyter>=1.1.1",
    "openai",
    "openai-agents>=0.6.5",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.2.1",
    "uvicorn>=0.34.2",
    "python-multipart>=0.0.18",
    "dateparser>=1.2.2",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "icalendar>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "mypy>=1.10.0",
    "ruff>=0.9.4",
    "types-dateparser>=1.2.2.20250809",
    "types-icalendar>=6.3.2.20251222",
]
